No 3D Array
Accomplished:
Packages are placed based off the known locations of other packages in
the trailer already, removing the need for a 3D array. Trailer can now
be incremented by inches, feet, centimeters, or any size without taking
up a large amount of memory. Only needed is the dimensions of the
trailer and the dimensions of the package.

Adds the dynamic ability to take the best package placement which fills
the trailer the most based off running the program with different
collision values. Found there is not a need for dynamic placement into
the trailer. The pallets should be placed against the back wall first
and not in the middle. The main goal is to list how to trailer would be
filled best based off of looking at the known collision values.

Adds the ability to export to a text file. The exported information is
what would be output to the console normally.


Still needs work:
Text file needs to export to a different location other than the project
folder. Look and see if the best placement of the algorithm is working
correctly.