Fixing Placement
Places items flush against the back wall of the trailer. Database function (findFragility) created in SeniorProject.cpp.

Needs database retrieval and maybe dynamically moving packages around after being placed. Finding a better location means that the package location is found and then the package is closer to the driver corner than before. Move placement after placing a package in the trailer. Look at all the packages placed and see if there is a better placement comparing the current locations of each. Needs to take in consideration weight and fragility.